Moved `key-generator.js` to `utils/key-generator.ts`
